pyproject.toml
[tool.poetry]
name = "fastapi-demo"
version = "0.1.0"
description = ""
authors = ["250304s <250304s@ugs.kochi-tech.ac.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = "^2.0.23"
aiomysql = "^0.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.21.1"
aiosqlite = "^0.19.0"
httpx = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

